Split new tests in new file
