+++ tests/unit/polars/test_write.py
@@ -1,5 +1,4 @@
 import datetime as dt
-import re
 from typing import Any
 from zoneinfo import ZoneInfo
 
@@ -349,194 +348,6 @@ def test_sink_parquet_partition_year(
     assert_frame_equal(lf, lf_roundtrip, check_row_order=False)
 
 
-def test_sink_parquet_partition_bucket(
-    shared_ducklake: dl.Ducklake, random_table_name: str
-) -> None:
-    # Arrange
-    # DuckLake's `bucket` transform is documented to use an Iceberg-compatible hash (32-bit
-    # Murmur3, x86 variant, seeded with 0), so the expected bucket ids below are derived from the
-    # official test vector published in the Iceberg spec's "32-bit Hash Requirements" appendix:
-    #   `hashLong(long(34)) == 2017239379`, so `bucket(8, 34) == 2017239379 % 8 == 3`.
-    # `35` and `36` were picked because they happen to land in different buckets from `34` and
-    # from each other, while `37` collides with `36`, so the test exercises both grouping and
-    # separation of partitions.
-    table = shared_ducklake.create_table(
-        random_table_name,
-        {"x": dl.Int64()},
-        partition_by=dl.PartitionColumn("x", transform="bucket", num_buckets=8),
-    )
-    lf = pl.LazyFrame({"x": [34, 35, 36, 37]})
-
-    # Act
-    table.sink_polars(lf)
-    lf_roundtrip = table.scan_polars()
-
-    # Assert
-    scan_result = table.scan()
-    assert len(scan_result.data_files) == 3
-    assert {_partition_value_from_path(f.path, "x") for f in scan_result.data_files} == {3, 5, 6}
-    assert_frame_equal(lf, lf_roundtrip, check_row_order=False)
-
-
-# Verified against a live Ducklake setup
-@pytest.mark.parametrize(
-    ("dtype", "pl_dtype", "value", "expected_bucket"),
-    [
-        pytest.param(dl.Varchar(), pl.Utf8(), "iceberg", 1, id="string"),
-        pytest.param(dl.Blob(), pl.Binary(), bytes([0, 1, 2, 3]), 1, id="binary"),
-        pytest.param(dl.Boolean(), pl.Boolean(), True, 4, id="boolean"),
-        pytest.param(dl.Date(), pl.Date(), dt.date(2017, 11, 16), 2, id="date"),
-        pytest.param(
-            dl.Timestamp("microseconds"),
-            pl.Datetime("us"),
-            dt.datetime(2017, 11, 16, 22, 31, 8),
-            7,
-            id="timestamp",
-        ),
-        pytest.param(dl.Float64(), pl.Float64(), 1.0, 7, id="float64"),
-    ],
-)
-def test_sink_parquet_partition_bucket_matches_iceberg_test_vectors(
-    shared_ducklake: dl.Ducklake,
-    random_table_name: str,
-    dtype: dl.DataType,
-    pl_dtype: pl.DataType,
-    value: object,
-    expected_bucket: int,
-) -> None:
-    # Arrange
-    table = shared_ducklake.create_table(
-        random_table_name,
-        {"x": dtype},
-        partition_by=dl.PartitionColumn("x", transform="bucket", num_buckets=8),
-    )
-    lf = pl.LazyFrame({"x": [value]}, schema={"x": pl_dtype})
-
-    # Act
-    table.sink_polars(lf)
-    lf_roundtrip = table.scan_polars()
-
-    # Assert
-    scan_result = table.scan()
-    assert len(scan_result.data_files) == 1
-    assert _partition_value_from_path(scan_result.data_files[0].path, "x") == expected_bucket
-    assert_frame_equal(lf, lf_roundtrip, check_row_order=False)
-
-
-def test_sink_parquet_partition_bucket_negative_value(
-    shared_ducklake: dl.Ducklake, random_table_name: str
-) -> None:
-    # Arrange
-    # The Iceberg spec doesn't publish a test vector for negative inputs, so this doesn't assert
-    # against an externally verifiable bucket id. Instead, it guards against a sign-handling
-    # regression: the mask applied to the hash (`& 0x7FFFFFFF`) must always clear the sign bit, so
-    # the resulting bucket id must be a valid, non-negative bucket regardless of whether the
-    # *input* value is negative, and the original value must still round-trip correctly.
-    table = shared_ducklake.create_table(
-        random_table_name,
-        {"x": dl.Int64()},
-        partition_by=dl.PartitionColumn("x", transform="bucket", num_buckets=8),
-    )
-    lf = pl.LazyFrame({"x": [-34]})
-
-    # Act
-    table.sink_polars(lf)
-    lf_roundtrip = table.scan_polars()
-
-    # Assert
-    scan_result = table.scan()
-    assert len(scan_result.data_files) == 1
-    assert 0 <= _partition_value_from_path(scan_result.data_files[0].path, "x") < 8
-    assert_frame_equal(lf, lf_roundtrip, check_row_order=False)
-
-
-def test_sink_parquet_partition_bucket_rejects_unverified_dtype(
-    shared_ducklake: dl.Ducklake, random_table_name: str
-) -> None:
-    # Arrange
-    # `Decimal`'s hash in the real DuckLake extension doesn't even match the value published in
-    # the Iceberg spec it's supposed to be compatible with (see `test_create_bucket_partition_
-    # rejects_unverified_dtypes` in `test_sink.py`), so bucketing on it must fail loudly rather
-    # than silently produce an incompatible bucket id.
-    table = shared_ducklake.create_table(
-        random_table_name,
-        {"x": dl.Decimal(9, 2)},
-        partition_by=dl.PartitionColumn("x", transform="bucket", num_buckets=8),
-    )
-    lf = pl.LazyFrame({"x": [1.0]}, schema={"x": pl.Decimal(9, 2)})
-
-    # Act & Assert
-    with pytest.raises(NotImplementedError, match="bucket"):
-        table.sink_polars(lf)
-
-
-def test_sink_parquet_partition_bucket_millisecond_timestamp(
-    shared_ducklake: dl.Ducklake, random_table_name: str
-) -> None:
-    # Arrange
-    # DuckLake hashes a `Timestamp` column's raw physical integer without converting between time
-    # units, so a millisecond-precision column hashes *differently* from the microsecond one in
-    # `test_sink_parquet_partition_bucket_matches_iceberg_test_vectors` for the same logical
-    # instant. Verified empirically against a real DuckDB + `ducklake` extension instance
-    # (`2017-11-16T22:31:08::TIMESTAMP_MS` hashes to `-1171784138`; masked and modulo `8` gives
-    # bucket `6`).
-    table = shared_ducklake.create_table(
-        random_table_name,
-        {"x": dl.Timestamp("milliseconds")},
-        partition_by=dl.PartitionColumn("x", transform="bucket", num_buckets=8),
-    )
-    lf = pl.LazyFrame(
-        {"x": [dt.datetime(2017, 11, 16, 22, 31, 8)]}, schema={"x": pl.Datetime("ms")}
-    )
-
-    # Act
-    table.sink_polars(lf)
-    lf_roundtrip = table.scan_polars()
-
-    # Assert
-    scan_result = table.scan()
-    assert len(scan_result.data_files) == 1
-    assert _partition_value_from_path(scan_result.data_files[0].path, "x") == 6
-    assert_frame_equal(lf, lf_roundtrip, check_row_order=False)
-
-
-@pytest.mark.parametrize(
-    ("dtype", "pl_dtype"),
-    [
-        (dl.Int8(), pl.Int8()),
-        (dl.Int16(), pl.Int16()),
-        (dl.Int32(), pl.Int32()),
-        (dl.Int64(), pl.Int64()),
-    ],
-    ids=lambda d: type(d).__name__,
-)
-def test_sink_parquet_partition_bucket_matches_across_integer_widths(
-    shared_ducklake: dl.Ducklake,
-    random_table_name: str,
-    dtype: dl.DataType,
-    pl_dtype: pl.DataType,
-) -> None:
-    # Arrange
-    # The Iceberg spec requires that "Integer and long hash results must be identical for all
-    # integer values" so that widening an integer column does not change its bucket assignment.
-    # `34` always hashes to bucket `3` out of `8` (see `test_sink_parquet_partition_bucket`),
-    # regardless of the physical width of the column being bucketed.
-    table = shared_ducklake.create_table(
-        random_table_name,
-        {"x": dtype},
-        partition_by=dl.PartitionColumn("x", transform="bucket", num_buckets=8),
-    )
-    lf = pl.LazyFrame({"x": [34]}, schema={"x": pl_dtype})
-
-    # Act
-    table.sink_polars(lf)
-
-    # Assert
-    scan_result = table.scan()
-    assert len(scan_result.data_files) == 1
-    assert _partition_value_from_path(scan_result.data_files[0].path, "x") == 3
-
-
 # ------------------------------------------- DEFAULTS ------------------------------------------ #
 
 
@@ -707,12 +518,6 @@ def test_sink_many_tiny_files(shared_ducklake: dl.Ducklake, random_table_name: s
 # ----------------------------------------------------------------------------------------------- #
 #                                              UTILS                                              #
 # ----------------------------------------------------------------------------------------------- #
-
-
-def _partition_value_from_path(path: str, column: str) -> int:
-    match = re.search(rf"/{column}=([^/]+)/", path)
-    assert match is not None, f"no `{column}=` partition segment found in path {path!r}"
-    return int(match.group(1))
 
 
 def read_table_stats(url: str, table: str) -> dict[str, Any]:
